refactor(pswap-chain): address remaining inline review comments
Per inline review comments on PR #8, addresses 10 of the 11 open
discovery.rs / observer.rs / lineage.rs / mod.rs / sqlite-store
findings in one cross-file pass (the iteration-shape comment on
discovery.rs:126 is still under design discussion).

Type narrowing
==============
* `current_depth`, `round_depth`, `PswapChainNoteUpdate::depth`:
  u64 -> u32 (review: "it is u32", "need to be u32",
  "Won't need this check when u32 is there"). The protocol's
  `PswapNoteAttachment::depth()` is already u32; narrowing matches.

* `PswapChainNoteUpdate::amount`: u64 -> AssetAmount (review:
  "Can we do this current.remaining_offered - remainder_cand.amount
  ...both in AssetAmount"). The observer converts at the boundary;
  the correlator now uses typed AssetAmount subtraction directly.

* All three `u32::try_from(round_depth)` checks (in `build_round_update`,
  `find_payback_index`, `reconstruct_payback`, and `mod.rs`) deleted
  along with the matching `Reconstruction(NoteError::other("...does
  not fit in u32"))` error paths.

Naming (discovery.rs)
=====================
* `current` (param + local) -> `lineage` (review: "current is also
  wrong usage of word")
* `block` (param) -> `at_block_num` (review: "Use the word block
  number not the block")
* `matches` (param + local) -> `notes` (review: "Don't use the word
  matches.....they are notes right?")
* `updates_by_order_depth` -> `notes_by_order_depth`
* `nullifier_blocks` -> `nullifier_to_block` (was misleading — reads
  like "blocks of nullifiers")

Tag-based payback/remainder identification
==========================================
Review: "We can differentiate them both based on the tag.......
please check it."

* `PswapChainNoteUpdate` gains a `tag: NoteTag` field, populated by
  the observer from `committed_note.metadata().tag()`. The correlator
  partitions the 2-note case using `original.storage().payback_note_tag()`:
  the payback note carries the creator-derived P2ID tag; the remainder
  carries the PSWAP asset-pair tag. No more try-each-as-payback
  reconstruction loop.

* `find_payback_index` (which iterated every candidate trying it as
  the payback) is deleted entirely. `reconstruct_payback` shrinks to
  ~13 LOC: build attachment with the now-u32 depth, reconstruct, verify
  id.

Removed defensive code
======================
* Consumer-must-match check in the partial-fill branch deleted
  (review: "We don't need this check....it is explicit... For now
  we can remove this"). The two notes are emitted by a single
  transaction; same sender is a protocol guarantee.

* The `n => /* 3+ candidates */ { error!(); Ok(None) }` arm replaced
  with `_ => unreachable!()` (review: "This won't be required....").
  The PSWAP script emits exactly 0, 1, or 2 notes per (order_id,
  depth) — 3+ is impossible by protocol. Companion unit test that
  asserted the old soft-skip behavior is removed.

Comment trims (discovery.rs)
============================
* `discover_pswap_rounds` doc preamble: 23 lines -> 6 (review:
  "All the comments are very big").
* 0-match (reclaim) branch: 9-line "outputs / cancel branch / signal
  alone does not tell us" walkthrough -> 2 lines (review on line 166:
  "Only creator of the note can consume it, so we know consumed it").
* Inner same-block multi-fill comment: 6 lines -> 3.
* "Saturating subtraction (preserves the v1 semantics of round off
  any over-fill rather than refuse to apply)" 4-line preamble in
  both branches -> 1 line in the full-fill branch only.
* Observer's `PswapChainNoteUpdate` doc: removed the "Naming follows
  the codebase convention" justification paragraph + per-field "the
  correlator decides via reconstruction" comments now superseded by
  the tag-based scheme.

Tests
=====
* `chain_update_from` test helper: now takes `depth: u32`, converts
  `amount: u64 -> AssetAmount` at the boundary, and threads
  `tag = note.metadata().tag()` so the correlator's tag-based
  payback/remainder identification works in tests.
* `pswap_attachment` test helper: `depth: u32` (no try_from needed).
* `build_round_update_more_than_two_candidates_returns_none` test
  removed (it tested the deleted soft-skip behavior).

All 17 PSWAP unit tests + 4 sqlite-store PSWAP tests pass clean.

+++ crates/rust-client/src/pswap/lineage.rs
@@ -86,8 +86,9 @@ pub struct PswapLineageRecord {
     /// observer can match incoming consumed-nullifier events against active
     /// lineages with a single query.
     pub current_tip_nullifier: Nullifier,
-    /// 0 for the original tip; increments by 1 each round.
-    pub current_depth: u64,
+    /// 0 for the original tip; increments by 1 each round. Matches the
+    /// protocol's `PswapNoteAttachment::depth()` (u32).
+    pub current_depth: u32,
     /// Offered-asset units still unfilled at this point in the chain. Starts
     /// at `original_pswap.offered_asset().amount()` and decreases by the
     /// per-round `payout_amount` until reaching zero. Typed as
@@ -205,7 +206,7 @@ pub struct PswapLineageRoundUpdate {
     pub order_id: Felt,
     /// `previous_depth + 1`. The protocol's PSWAP script stamps this in the
     /// attachment word of every output note emitted in this round.
-    pub round_depth: u64,
+    pub round_depth: u32,
     /// Account that consumed the previous tip and emitted the new outputs.
     /// For a reclaim, equals the creator.
     pub consumer_account_id: AccountId,
@@ -282,7 +283,7 @@ pub fn build_record_from_columns(
     original_pswap: PswapNote,
     current_tip_note_id: NoteId,
     current_tip_nullifier: Nullifier,
-    current_depth: u64,
+    current_depth: u32,
     remaining_offered: u64,
     remaining_requested: u64,
     last_consumer_account_id: Option<AccountId>,
